requirements.txt
beautifulsoup4==4.11.2
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-functions==1.10.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-http-client==3.3.7
requests==2.28.2
rsa==4.9
sendgrid==6.9.7
six==1.16.0
soupsieve==2.4
starkbank-ecdsa==2.2.0
urllib3==1.26.14
functions-framework==3.*
